reset password

+++ server/controllers/reset/newPassword.js
@@ -1,27 +1,47 @@
-import Token from "../../models/token.js";
 import Users from "../../models/user.js";
+import Token from "../../models/token.js";
+import bcrypt from "bcrypt";
 
 export const newPassword = async (req, res) => {
-  const {id, token} = req.params
+  const { id, token, password } = req.body;
+  console.log('id', id)
+  console.log('token', token)
+  console.log('password', password)
 
   try {
-    const user = await Users.findOne({_id: id });
+    const user = await Users.findOne({ _id: id });
     if (!user) {
-      res.json("invalid link");
+      return res.status(400).json("Invalid link: User not found");
     }
 
     const verificationToken = await Token.findOne({
       userId: user._id,
-      token: token
+      token: token,
     });
 
     if (!verificationToken) {
-      res.json("invalid link");
+      return res.status(400).json("Invalid link: Token not found");
     }
-    await Users.updateOne({ _id: user._id }, { password: req.body.password });
+
+    // Add logging to check the value of req.body.password
+    console.log("New password:", password);
+
+    // Hash the new password before updating
+    const hashedPassword = await bcrypt.hash(req.body.password, 10);
+
+    // Always set the password to the hashed password
+    user.password = hashedPassword;
+
+    const updated = await user.save();
+
+    if (!updated) {
+      return res.status(500).json("Error updating password");
+    }
+
     await Token.deleteOne({ _id: verificationToken._id });
-    res.json("password updated");
+    return res.json("Password updated");
   } catch (error) {
-    res.status(500).json("Error occured")
+    console.error("Error:", error);
+    return res.status(500).json("Error occurred");
   }
 };
